fix(simd-bench): drop redundant i32 cast in the aarch64-only shannon_entropy setup — un-breaks Clippy (aarch64)
Was masked by the dead_code error the previous commit fixed (clippy
stops at the first deny); only the arm job compiles this cfg body.
Verified via cargo clippy --target aarch64-apple-darwin -p
jxl-encoder-simd --all-targets -- -D warnings.

+++ jxl-encoder-simd/benches/kernel_tiers.rs
@@ -169,7 +169,7 @@ fn bench_kernels(suite: &mut Suite) {
     {
         let counts: &'static [i32] = Box::leak(
             (0..4096)
-                .map(|i| ((i * 7919) % 997) as i32)
+                .map(|i| (i * 7919) % 997)
                 .collect::<Vec<i32>>()
                 .into_boxed_slice(),
         );
